implement quotes count function
