+++ src/lexer/lexer_utils.c
@@ -67,49 +67,4 @@ void	set_type(t_token *new)
 	new->type = syntax;
 }
 
-int	is_quoted(char *str, size_t i, char token)
-{
-	char	cmp;
-	size_t	j;
-	int		sv;
 
-	j = i;
-	sv = 0;
-	cmp = '\'';
-	if (token == '\\')
-		return (0);
-	if (token == '\'')
-		cmp = '\"';
-	while (str[j] && str[j] != cmp)
-		if (str[++j] == cmp)
-			sv = 1;
-	j = i;
-	while (j && str[j] != cmp)
-		if (str[--j] == cmp && sv)
-			return (1);
-	return (0);
-}
-
-char	*ft_decrustify_str(char *str)
-{
-	size_t	i;
-	size_t	j;
-	char	*res;
-	char	*tmp;
-
-	i = 0;
-	j = 0;
-	tmp = ft_calloc(ft_strlen(str) + 1, sizeof(char *));
-	while (str[j])
-	{
-		if (str[j] != '\"' && str[j] != '\'' && str[j] != '\\')
-			tmp[i++] = str[j++];
-		else if (j > 0 && (str[j - 1] == '\\' || is_quoted(str, j, str[j])))
-			tmp[i++] = str[j++];
-		else
-			j++;
-	}
-	res = ft_calloc(i + 1, sizeof(char));
-	ft_strlcpy(res, tmp, i + 1);
-	return (free(str), free(tmp), res);
-}
